Move messages closer to spot of usage
