+++ front/app/containers/Admin/projects/components/ProjectRow/AdminTag.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { colors } from '@citizenlab/cl2-component-library';
 import { StyledStatusLabel } from '.';
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from '../../messages';
+import messages from '../messages';
 import Link from 'utils/cl-router/Link';
 import { adminProjectsProjectPath } from 'containers/Admin/projects/routes';
