+++ front/app/containers/Admin/projects/components/ProjectRow/ManageButton.tsx
@@ -3,7 +3,7 @@ import { adminProjectsProjectPath } from 'containers/Admin/projects/routes';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from '../../messages';
+import messages from '../messages';
 
 // components
 import { RowButton } from '../StyledComponents';
