+++ front/app/containers/Admin/projects/components/messages.ts
@@ -32,4 +32,17 @@ export default defineMessages({
     id: 'app.containers.Admin.projects.all.copyProjectButton',
     defaultMessage: 'Copy project',
   },
+  xGroupsHaveAccess: {
+    id: 'app.containers.AdminPage.ProjectEdit.xGroupsHaveAccess',
+    defaultMessage:
+      '{groupCount, plural, no {# groups can view} one {# group can view} other {# groups can view}}',
+  },
+  onlyAdminsCanView: {
+    id: 'app.containers.AdminPage.ProjectEdit.onlyAdminsCanView',
+    defaultMessage: 'Only admins can view',
+  },
+  editButtonLabel: {
+    id: 'app.containers.AdminPage.ProjectEdit.editButtonLabel',
+    defaultMessage: 'Edit',
+  },
 });
